+++ harbor_cli/config.py
@@ -293,16 +293,20 @@ class OutputSettings(BaseModel):
     format: OutputFormat = OutputFormat.TABLE
     paging: bool = Field(
         False,
-        description="Show output in pager (if supported). Does not support color output currently.",
+        description="Show output in pager (if supported). Default pager does not support color output currently.",
     )
-    # Custom pager support in Rich doesn't seem very mature out of the box,
-    # and it would require us to write a custom pager class to support it.
     pager: Optional[str] = Field(None, description="Pager to use if paging is enabled.")
-
     # Naming: Don't shadow the built-in .json() method
     # The config file can still use the key "json" because of the alias
     table: TableSettings = Field(default_factory=TableSettings)
     JSON: JSONSettings = Field(default_factory=JSONSettings, alias="json")
+    confirm_enumeration: bool = Field(
+        True,
+        description=(
+            "Show confirmation prompt for certain resource enumeration "
+            "commands when invoked without a limit or filter. E.g. `auditlog list`"
+        ),
+    )
 
     @validator("pager")
     def set_pager(cls, v: Optional[str]) -> Optional[str]: